Code style and cleanup
