+++ app/src/test/java/com/mysugr/android/testing/example/net/BackendGatewayMockSteps.kt
@@ -11,7 +11,7 @@ import org.mockito.Mockito.verify
 
 class BackendGatewayMockSteps(testContext: TestContext) : BaseSteps(testContext, appModuleTestingConfiguration) {
 
-    override fun configure() = super.configure().requireMock<BackendGateway>().onSetUp(this::setUp)
+    override fun configure() = super.configure().requireMock<BackendGateway>().onSetUp(::setUp)
 
     private val instance by dependency<BackendGateway>()
     private var backendUsers = mutableListOf<BackendFakeUser>()
