+++ app/src/test/java/com/mysugr/android/testing/example/view/LoginSteps2.kt
@@ -1,7 +1,7 @@
 package com.mysugr.android.testing.example.view
 
 import com.mysugr.android.testing.example.appModuleTestingConfiguration
-import com.mysugr.android.testing.example.auth.AuthManagerSteps
+import com.mysugr.android.testing.example.auth.AuthManagerSteps2
 import com.mysugr.android.testing.example.net.BackendFakeSteps
 import com.mysugr.android.testing.example.net.BackendGateway
 import com.mysugr.sweetest.framework.base.BaseSteps
@@ -21,7 +21,7 @@ class LoginSteps2(testContext: TestContext) :
 
     override fun configure() = super.configure()
         .requireReal<LoginViewModel>()
-        .requireSteps<AuthManagerSteps>()
+        .requireSteps<AuthManagerSteps2>()
 
     fun whenLoggingIn(email: String, password: String) {
         viewModel.loginOrRegister(email, password)
